+++ examples/resources/fmc_external_certificate/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_external_certificate.example "<id>"