+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.0.0-rc5 (Unreleased)
 
+- (Enhancement) Add support for `fmc_internal_certificate`, `fmc_internal_certificate_authority`, `fmc_external_certificate`, `fmc_trusted_certificate_authority` resources and data sources
+- (Enhancement) Add support for `fmc_certificate_enrollment` resource and data source
 - (Enhancement) `fmc_access_control_policy` has new attributes `manage_rules` and `manage_categories` that disable managing (resource) or reading (data source) rules and categories
 - (Enhancement) Add support for `fmc_access_rule` resource and data source
 - (Enhancement) (Early access) Add support for `fmc_access_rules` resource and data source